Avoid extra allocation+sorting+deduping in `dkg::round1()`
Currently `dkg::round1()` allocates a vector for the list of
participants, and then sorts and dedupes this vector. It does so twice:
once for calculating `max_signers`, once for calculating the `checksum`.

This commit changes the behavior so that the vector allocation, sorting,
and deduping happens only once and the result is used for both purposes.