Make features requiring `std` optional
This commit changes some of the functionality of this crate to not use
`std` or `alloc`.

It also adds the following optional features to this crate:
- `std`
- `signing`
- `dkg`

All the features except `dkg` are enabled by default and may be disabled
with `default-features = false` in `Cargo.toml`.

The main goal of this commit is to allow `no_std` creates to consume
a subset of this crate.