implements minimized round2 version
round2_min follows the example of round3_min and inputs/outputs the minimal
amount of data to support dkg

takes as input:
- a participant secret
- a vector of byte arrays, one for each participant identity (excluding the self
  identity)
- a byte array containing the encrypted round1 secret package
- a vector of round1 frost public packages, one for each participant (excluding
  the user's own public package)
- an rng

round2_min returns the encrypted round2 secret package and a vector of round2
frost public packages. it does NOT create a CombinedPublicPackage or ironfish
PublicPackages from the frost public packages. the user can create these
separately (e.g., on the host machine instead of a ledger device) before
distributing the packages.

the vector of round2 public packages is returned in order of the participants
passed in: round2_frost_packages[i] is intended for participant[i]