removes signature from SigningCommitment
our SigningCommitment type includes a signature of the identity of the
participant creating the commitment, the raw signing commitments, and the input
checksum

the signature allows other participants to verify that the commitment was
created by the participant whose identity is included in the SigningCommitment.
this verification takes place before creating a SigningPackage. without this
verification an invalid SigningPackage can be created: the participant identity
with the mismatched signing commitment will be unable to generate a signature
share. the signature therefore allows the signing process to terminate one step
earlier in case a participant generates an invalid commitment.

the multisig Ledger app will not be able to generate this signature without
additional features.

these changes remove the signature from the SigningCommitment struct so that we
will be able to create SigningCommitment instances from raw commitments
generated on Ledger devices.